requirements.txt
fastapi
uvicorn
opencv-python
pillow
ultralytics
streamlit